backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-login==0.6.2
flask-cors==4.0.0
werkzeug==2.3.7
python-dotenv==1.0.0
openpyxl==3.1.2
psycopg2-binary==2.9.7
sqlalchemy==2.0.23